onCheck, onCall implemented.

+++ app/src/main/kotlin/me/buddha/chiplesspoker/data/model/PotEntity.kt
@@ -4,7 +4,7 @@ import me.buddha.chiplesspoker.domain.model.Pot
 
 data class PotEntity(
     val chips: Long,
-    val players: List<Long>,
+    val players: List<Int>,
 )
 
 fun PotEntity.asExternalModel() = Pot(
